r/types: moved `raft::vnode` to fundamentals header
(cherry picked from commit eaa840887e999d4b46fd566b6d3d1c0b660c0790)
