+++ src/v/raft/group_configuration.h
@@ -11,8 +11,8 @@
 
 #pragma once
 #include "model/metadata.h"
+#include "raft/fundamental.h"
 #include "reflection/adl.h"
-#include "serde/rw/envelope.h"
 #include "serde/rw/vector.h"
 
 #include <boost/range/join.hpp>
@@ -27,34 +27,7 @@ struct broker_revision {
 };
 
 static constexpr model::revision_id no_revision{};
-class vnode
-  : public serde::envelope<vnode, serde::version<0>, serde::compat_version<0>> {
-public:
-    constexpr vnode() = default;
 
-    constexpr vnode(model::node_id nid, model::revision_id rev)
-      : _node_id(nid)
-      , _revision(rev) {}
-
-    bool operator==(const vnode& other) const = default;
-    bool operator!=(const vnode& other) const = default;
-
-    friend std::ostream& operator<<(std::ostream& o, const vnode& r);
-
-    template<typename H>
-    friend H AbslHashValue(H h, const vnode& node) {
-        return H::combine(std::move(h), node._node_id, node._revision);
-    }
-
-    constexpr model::node_id id() const { return _node_id; }
-    constexpr model::revision_id revision() const { return _revision; }
-
-    auto serde_fields() { return std::tie(_node_id, _revision); }
-
-private:
-    model::node_id _node_id;
-    model::revision_id _revision;
-};
 /**
  * Enum describing configuration state.
  *
